Bar Chart Refactor Moved Bar chart to its own Swiftui File
