+++ SwiftDataTesting/Views/BarChartView.swift
@@ -5,4 +5,61 @@
 //  Created by Benia Morgan-Ware on 5/29/24.
 //
 
-import Foundation
+import SwiftUI
+import Charts
+
+
+struct BarChartView: View {
+    @State private var isAnimating = false
+    let groupedExpenses: TransactionGroup
+    
+    //Getting max amount spent in a month
+    //Defaults to 1000 is no values exist
+    var maxSpending: Double {
+        groupedExpenses.values.map {$0.sum }.max() ?? 1000
+    }
+    var minSpending: Double {
+        groupedExpenses.values.map {$0.sum }.min() ?? 50
+    }
+    
+    var body: some View {
+        Chart(groupedExpenses.keys, id: \.self) { month in
+            if let group = groupedExpenses[month]  {
+                
+                BarMark(x: .value("Month", month),
+                        y: .value("Total Spent", isAnimating == false ? 50 : group.sum)
+                )
+                .foregroundStyle(by: .value("Month", month))
+                
+            }
+            
+        }
+        .chartLegend(.hidden)
+        .chartForegroundStyleScale(range: [Color.DarkBlue, Color.brandGreen, Color.navyBlue, Color.lightblue])
+        .onAppear {
+            withAnimation {
+                isAnimating = true
+            }
+        }
+        .onDisappear {
+            isAnimating = false
+        }
+        .chartYScale(domain: minSpending...maxSpending)
+        //.padding(.bottom, 40)
+        .frame(height: 140)
+        
+        .chartXAxis {
+            AxisMarks(stroke: StrokeStyle(lineWidth: 0))
+            
+        }
+        
+        .chartYAxis {
+            AxisMarks(stroke: StrokeStyle(dash:[7]))
+        }
+        .listRowSeparator(.hidden)
+    }
+}
+
+#Preview {
+    BarChartView(groupedExpenses: [:])
+}
