+++ SwiftDataTesting/Views/ChartsView.swift
@@ -16,7 +16,7 @@ struct ChartsView: View {
     @State private var barGraphIsAnimating = false
     @State private var donutGraphIsAnimating = false
     @State private var barGraphPicker: TimePeriods = .year
- 
+    
     
     func getMonthlyExpenseSum() -> TransactionGroup {
         guard !expenses.isEmpty else { return [:] }
@@ -130,50 +130,15 @@ struct ChartsView: View {
                             }
                             .padding(.bottom, 25)
                             
-                            
-                            
                             //MARK: Bar Chart starts here
-                            Chart(groupedExpenses.keys, id: \.self) { month in
-                                if let group = groupedExpenses[month]  {
-                                    
-                                    BarMark(x: .value("Month", month),
-                                            y: .value("Total Spent", barGraphIsAnimating == false ? 50 : group.sum)
-                                    )
-                                    .foregroundStyle(by: .value("Month", month))
-                                    
-                                }
-                                
-                            }
-                            .chartLegend(.hidden)
-                            .chartForegroundStyleScale(range: [Color.DarkBlue, Color.brandGreen, Color.navyBlue, Color.lightblue])
-                            .onAppear {
-                                withAnimation {
-                                    barGraphIsAnimating = true
-                                }
-                            }
-                            .onDisappear {
-                                barGraphIsAnimating = false
-                            }
-                            .chartYScale(domain: 50...1000)
-                            //.padding(.bottom, 40)
-                            .frame(height: 140)
-                            
-                            .chartXAxis {
-                                AxisMarks(stroke: StrokeStyle(lineWidth: 0))
-                                
-                            }
-                            
-                            .chartYAxis {
-                                AxisMarks(stroke: StrokeStyle(dash:[7]))
-                            }
-                            .listRowSeparator(.hidden)
+                            BarChartView(groupedExpenses: groupedExpenses)
                             
                             
                             DisclosureGroup("Spending Insights") {
                                 VStack {
                                     HStack {
                                         Text("Top Spending by Category")
-                                
+                                        
                                     }
                                     //MARK: Donut Chart starts here
                                     DonutChartView(categoryTotals: topCategoryTotals)
@@ -182,11 +147,8 @@ struct ChartsView: View {
                                 
                             }
                             
-                            
-                            
                         }
                     }
-                    
                     
                     
                 }
@@ -199,63 +161,9 @@ struct ChartsView: View {
 }
 
 
-struct BarChartView: View {
-    @State private var isAnimating = false
-    let groupedExpenses: TransactionGroup
-    
-    //Getting max amount spent in a month
-    //Defaults to 1000 is no values exist
-    var maxSpending: Double {
-        groupedExpenses.values.map {$0.sum }.max() ?? 1000
-    }
-    var minSpending: Double {
-        groupedExpenses.values.map {$0.sum }.min() ?? 50
-    }
-    
-    var body: some View {
-        Chart(groupedExpenses.keys, id: \.self) { month in
-            if let group = groupedExpenses[month]  {
-                
-                BarMark(x: .value("Month", month),
-                        y: .value("Total Spent", isAnimating == false ? 50 : group.sum)
-                )
-                .foregroundStyle(by: .value("Month", month))
-                
-            }
-            
-        }
-        .chartLegend(.hidden)
-        .chartForegroundStyleScale(range: [Color.DarkBlue, Color.brandGreen, Color.navyBlue, Color.lightblue])
-        .onAppear {
-            withAnimation {
-                isAnimating = true
-            }
-        }
-        .onDisappear {
-            isAnimating = false
-        }
-        .chartYScale(domain: minSpending...maxSpending)
-        //.padding(.bottom, 40)
-        .frame(height: 140)
-        
-        .chartXAxis {
-            AxisMarks(stroke: StrokeStyle(lineWidth: 0))
-            
-        }
-        
-        .chartYAxis {
-            AxisMarks(stroke: StrokeStyle(dash:[7]))
-        }
-        .listRowSeparator(.hidden)
-    }
-}
-
-#Preview {
-    BarChartView(groupedExpenses: [:])
-}
-
-
 #Preview {
     let preview = previewContainer([Expense.self])
     return ChartsView().modelContainer(preview.container)
 }
+
+
